+++ src/main/java/lotto/domain/Answer.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain;
 
 public class Answer {
     private Lotto answerNumbers;
