+++ src/main/java/lotto/domain/Profit.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain;
 
 public class Profit {
     private final long integerPart;
