+++ src/main/java/lotto/enums/Ranking.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.enums;
 
 public enum Ranking {
     FAIL, FIFTH, FOURTH, THIRD, SECOND, FIRST
