+++ src/test/java/lotto/domain/ResultTest.java
@@ -1,5 +1,7 @@
-package lotto;
+package lotto.domain;
 
+import lotto.domain.*;
+import lotto.enums.Ranking;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
